# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Thalsecom eCommerce API",
        "description": "eCommerce API with JWT authentication, inventory management, product management, and order processing",
        "contact": {
            "email": "admin@thalsecom.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8001",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/v1/inventory/variants/{id}/logs": {
            "get": {
                "tags": [
                    "Inventory"
                ],
                "summary": "Get inventory logs for a variant",
                "operationId": "988fbb6d4c13acd0ac0158c5721d013d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Inventory logs retrieved"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/inventory/variants/{id}/add": {
            "post": {
                "tags": [
                    "Inventory"
                ],
                "summary": "Add stock to variant",
                "operationId": "be41dbcd7f4ba1186974bbccbc0f5240",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "quantity"
                                ],
                                "properties": {
                                    "quantity": {
                                        "type": "integer",
                                        "example": 50
                                    },
                                    "reason": {
                                        "type": "string",
                                        "example": "Restocked from supplier"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Stock added successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/inventory/variants/{id}/adjust": {
            "post": {
                "tags": [
                    "Inventory"
                ],
                "summary": "Adjust stock quantity",
                "operationId": "67f9747407dc263cdbbc579bce6d2222",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "new_quantity"
                                ],
                                "properties": {
                                    "new_quantity": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "reason": {
                                        "type": "string",
                                        "example": "Physical inventory count"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Stock adjusted successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/inventory/alerts": {
            "get": {
                "tags": [
                    "Inventory"
                ],
                "summary": "Get all low stock alerts",
                "operationId": "d5651196fc39fe2ac066530ed2d3d278",
                "responses": {
                    "200": {
                        "description": "Low stock alerts retrieved"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/inventory/alerts/{id}/resolve": {
            "post": {
                "tags": [
                    "Inventory"
                ],
                "summary": "Mark low stock alert as resolved",
                "operationId": "2f1dba72b7fa8c53b28cae64fe44f4d6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Alert resolved"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/inventory/stock-summary": {
            "get": {
                "tags": [
                    "Inventory"
                ],
                "summary": "Get inventory stock summary",
                "operationId": "630205ff042246d915ca1cd45896cc9d",
                "responses": {
                    "200": {
                        "description": "Stock summary retrieved"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}